LayoutTests/imported/w3c/web-platform-tests/html/canvas/offscreen/text/2d.text.measure.baselines.worker-expected.txt


FAIL Testing baselines assert_equals: ctx.measureText('A').ideographicBaseline === 6.25 (got -39[number], expected 6.25[number]) expected 6.25 but got -39
